// cloudapp/src/app/models/response.ts
export class PriceResponse {
    requestId: string = "";
    availablity: string = "";
    errorMsg: string = "";
    price: string = "";
    ginPublisherCode: string = "";
  }

  export class PlaceOrderResponse {
    requestId: string = "";
    status: string = "";
    errorMsg: string = "";
    randomNumber: string = "";
    additionalId: string = "";
  }